requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
idna==3.3
insta-scrape==2.1.2
Pillow==9.0.1
requests==2.27.1
soupsieve==2.3.1
tk==0.1.0
urllib3==1.26.8
